3_Ensemble/requirements.txt
scanpy==1.11.4
anndata==0.12.2
numpy==2.2.6
pandas==2.3.2
scikit-learn==1.7.2
xgboost==3.0.5
lightgbm==4.6.0
scvi-tools==1.4.0
celltypist==1.7.1
torch==2.8.0
lightning==2.5.5
matplotlib==3.10.6
seaborn==0.13.2
openpyxl==3.1.5
